requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-ckeditor==6.1.0
django-js-asset==1.2.2
idna==3.2
Pillow==8.0.0
pycodestyle==2.7.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
